mid_registrar: Skip usrloc callbacks for foreign domains
In case mid_registrar is used alongside registrar using separate domains
(location tables), then mid_registrar must not process any callbacks for
Contacts/AORs which it is not responsible for.

A similar logic should be added for registrar as well, however it
currently does not subscribe to any usrloc callbacks.

Fixes #2716