PN PURR processing: Handle "no Route headers" as success
If the ";pn-purr=" parameter is not present in the R-URI and the request
has no Route headers to search for it, it is a non-PN request.  So
return a "success" return code (2), as there is nothing left to do here.

(cherry picked from commit b1fed3e74ea3529f6b1190f8104bf4181d28ec9a)