dialplan: Always set the "attrs" pvar on success; Improve docs
This patch avoids leaving an untouched "attrs_pvar" variable after a
successful dp_translate(), in case the attributes are NULL or
empty-string.

In order to both prevent scripting bugs as well as boilerplate
initialization code, the "attrs_pvar" is now always set to ""
(empty-string) in case the "attrs" field is either NULL or empty-string.