Add the "tcp_mgm" module
This module is entirely optional, and it offers a DB-based way of
customizing the behavior of certain TCP connections depending on the
protocol employed (e.g. BIN vs. TCP vs. MSRP, etc.), or specific
combinations of remote/local IP addresses.

When the module is not loaded, or when some connections fail to match
any of the DB-defined TCP paths, the default TCP connection profile will
be used instead, thus OpenSIPS remains fully backwards-compatible with
the previous version.

See 0aa8e9514e for more info on TCP connection profiles.