Improve the TCP conns balancing over the TCP workers
Instead of balancing as number of conns per worker, balance based on the internla load of the TCP workers.
At the end we want to avoid overloading some TCP workers, so to have a even load distribution. The number of TCP conns is irrelevant (as they may have different amount of traffic which may translate in different load values).