msrp_gateway: properly clean up lingering sessions
Add a session timeout after which to end a session if no message has
been received from either side for that session. Also, add a specific
timeout that is triggered by no activity on the MESSAGE side.

Also add the ability to list and end sessions from the outside, via MI
functions.