[db_oracle] Do not change asynch mode when connection lost.
The sessions may be ended due to a network error,
and change OCI_ATTR_NONBLOCKING_MODE attribute with no session
will cause "ORA-03126: network driver does not support non-blocking
operations" error.

In the case, the function done_timelimit will failed with ORA-03126 error.
The error is not considered as a connection loss,
and subsequent operations will always throw this error,
and the lost connection never reconnect.