Fix problem where incorrect NOTIFYs are sent
When a presentity expires, opensips will clean it correctly. In case of
dialog event presentities, it is necessary to notify that the dialog
event has expired.

To do so, opensips 3 will generate a NOTIFY from the DB as the original
NOTIFY request does not have a body resulting from a PUBLISH. In that
case, we had two NOTIFY requests following up eachother, one with an
incorrect old state/body caused by the old presentity that was expiring.

This patch fixes that problem.