Fix in-dialog ACK being relayed with a wrong CSEQ
There are cases where opensips can handle a reINVITE before the ACK that
acknowledges the previous INVITE/200 OK exchange.

In that case, the ACK would be relayed with the reINVITE CSEQ, leading
to interoperability issues.

We now never use the generated invite cseq for ACK requests. This change
also works with reINVITE pings.

It seems the problem was introduced in
27b0b21b2f2aa7b29fdaf0b7bc35274979fc55b0

Closes https://github.com/OpenSIPS/opensips/issues/1071