b2b_entities: properly handle negative replies to reINVITEs
Do not change the dialog state to terminated as a result of a sending a
negative reply to a reINVITE. This would have prevented the forwarding of
further requests.

Thanks to David Escartin from Sonoc for reporting and testing the fix.

(cherry picked from commit e8df85a64a3b8095cd95c0b2d9254de605b61384)