b2b_logic: prevent possible crashes when accessing $b2b_logic vars
Make sure to always search in the hash for the current tuple instead of relying
on a pointer saved in the b2b context. There might be races between processing
BYEs and deleting the tuple and running a b2b_logic reply route where we access
a $b2b_logic variable.

Thanks to David Escartin from Sonoc for reporting.