Fix profile sizes count on restart
In case of OpenSIPS crash or reload, profile sizes were incremented on
reload when the dialog module was reloading the ongoing dialogs from the
database.

In case a cache module is used to store profile values, we might expect
it to be correct when OpenSIPS crashes. If we increment the profile size
when OpenSIPS restarts, then the counters are incremented twice for the
same dialog.