tls_wolfssl: fix differing TLS version for write duplicate struct
When using a range of supported TLS versions, the WOLFSSL write duplicate
structure uses the highest TLS version set for the context instead of the one
actually negociated for the current session. This may lead to errors for TLS
write or shutdown operations, for ciphers that are not supported in the
TLS version used by the write dup structure.

Fixes #2729

(cherry picked from commit 35b29e618b1175618c61747591dcf87370056102)