clusterer: add ability to disable replication packets dispatching
Add a new "dispatch_jobs" modparam which controls whether the processing of
replicated packets is dispatched through IPC or not.
Disabling the dispatching mechanism prevents high CPU loads caused by
the "thundering herd" problem.

Credits go to Rizwan Syed and Connex Carrier Services for supporting the
troubleshooting of this issue.