clusterer: Always broadcast a packet after shtag activation
Due to network outages, it is currently possible for both sharing tags
of an Active/Backup pair to end up in ACTIVE state.  This patch makes it
so a "cluster_shtag_set_active" MI command always broadcasts (forces)
the ACTIVE state onto all neighbours, thus fixing the incorrect,
post-outage state of the tags.

Credits to @bogdan-iancu and @rvlad-patrascu for their contributions to
this fix!