[dialog] fix bad unref leading to dialog leaking
When using the dlg_on_timeout to set a new timeout for the dialog, due to the timer related operations (re-inserting in the list), an extra ref cnt happens. If not unref'ed, the dialog will stay (as terminated) forever in memory.

Closes #2788

(cherry picked from commit 4062de45d0ccfe62bb4f9475f6320d2726349bab)