[TCP] fix init TCP support for auto-scalled procs
When starting the TCP Main proc, be sure we initialize the MAIN<>Worker communication sockets also for the procs to be potentially scaled in the future.
`pt[n].unix_sock` is valid only for running procs (not for the "autoscale to-be"), so use the `pt[n].tcp_socks_holder[0]` holder (valid for all procs, running or not).
Thanks go to Yury Kirsanov and Slava Bendersky for reporting and helping with the troubleshooting.

(cherry picked from commit d7a4bf8f134d3d2cd761f528e6963445c9ef9b06)