b2b_entities: fix shm memory leak
For the write back db mode, the entity storage data was only freed after
inserting/updating the entity in the database. This would lead to memory
leaks if the entity was deleted before the DB timer task had a chance
to be run.

Thanks go to Norm Brandinger from Five9 for reporting!