tracer: Be more verbose when cannot start trace
Currently, when tracer cannot initiate a specified tracing mode it tries
to switch to another tracing mode emitting just a debug message. Debug
messages are relatively rarely being turned on, so this switch can be
left unnoticed. Let's raise the severity to WARN so this situation can
be spotted much easier.

Signed-off-by: Peter Lemenkov <lemenkov@gmail.com>