async(): Attempt to recover from abort() state
It seems the "if" condition at the start of t_resume_async() is
actually reachable under the right circumstances and causes an
immediate program crash.

As an intermediate improvement until the source of the issue is found,
try to simply overwrite the unexpected context, if one is found.  A
small memory leak coupled with a CRITICAL log message is preferable to
an instant, forced crash.