cgrates: fix infinite re-connect issue
When a non-default connection would have reconnected, due to a
mishandling of the return code of the `cgrc_reconn` function, a new
connection would have been created every time, resulting in infinite
re-connections. And in time, this would acquire all the available ports
(65k) and would no longer accept new connections.

Many thanks go to Vasilios Tzanoudakis (@vtzan on GitHub) for reporting
and providing a solid testbed for this.
Close #2850