mid_registrar: Fix occasional undesired De-REGISTER forwarding in mode 2
If the timings were aligned in a specific way, the mid-registrar (in AoR
throttling mode) would actually forward a De-REGISTER downstream even
though it had at least 1+ healthy remaining contacts after having
applied the De-REGISTER locally.  This would cause some degree of
service loss, as the remaining UAs were registered on the mid-reg, but
not on the backend.

Many thanks to Carlos Cruz Luengo (@cruzccl) and @ryan-netizen for an
accurate report, as well as helping come up with the fix!

Fixes #2615