tm: remove cluster received '\0' from buffer len
When sending a message through cluster, the tm module adds a '\0'
terminator to the message - that is because receive_msg expects it.
However, this character should not be part of the message lenght.

Reported by Denys Pozniak on mailing lists.

(cherry picked from commit 81e9b14a16acd284469d8958c57dcece69699a85)