b2b_entities: properly handle BYE while another request is pending
Instead of rejecting the BYE with 491, first complete the transactions on the
UAS side with 200 OK for the BYE and 487 for the other pending transaction.
Also, mark the tuple for deletion at b2b_logic level and send a BYE to
the peer entity after completing any UAC transcations on this side.

Credits to David Escartin from Sonoc for reporting.