Add support for in-memory logging of shm operations
This is enabled via the new 'shm_memlog_size' global parameter which
configures the maximum number of operations to keep in the history.
A separate memory block, dedicated for this shm debug info will be
allocated. A single 'struct struct_hist' type object will be added to the
list for each alloc/free operation.