registrar: Fix possible crash in remove_ip_port()
The release_urecord() function may actually delete the AoR on the spot,
including the map_t node which the iterator is currently at!  This was
causing an invalid SHM memory read when calling iterator_next(),
potentially running into a crash.

The fix is to advance the iterator ahead of time, before the
release_urecord() call.

Fixes #2954