uac: Add basic detection for multiple uac_replace_xxx() calls
This patch adds a lightweight detection of multiple calls to
uac_replace_from() or uac_replace_to() during the processing of a SIP
request, using an int bitmask stored in the current SIP message context.
Returning an error helps avoid producing a bad SIP From/To header, with
doubled URIs due to doubled lumps, etc.

Still, to compromise between safety and performance, the case where
multiple uac_replace_xxx() get called during a failure_route is not
detected, as this would require extra TMCB_ON_FAILURE callback, which is
a bit of overkill just for the sake of scripting QoL...

Fixes #2720