event_route: fix crash when raising with no params
When raising an event using `raise_event` with AVP values that are not
correctly populated, it triggers a corner case in the event_route
module: the module "serializes" the parameters just after the event,
hence, if there is no expanded parameter, the first parameter ends up
pointing to a bogus value, resulting in a crash.

This commit fixes this corner case by treating  it (0 expanded
parameters) just as if there no parameters passed at all.

Many thanks to David Escartin from Sonoc for the help in troubleshooting
this issue!