b2b_logic: ignore replies to NOTIFYs sent while bridging
This will get rid of the races between the processing of the replies for the
NOTIFY and the subsequent BYE sent to the bridge initiator. The entity
might be already deleted as a result of handling the 200 OK for the BYE
before we try to do a lookup again for handling the NOTIFY reply. These
races would not produce any significant negative effects but OpenSIPS
would log unnecessary 'No b2b_key match found' errors.

(cherry picked from commit a72cfbeda919eca0020d34e33c844cc3b49a3b71)