b2b_logic: fix bridging with provisional media and rollback feature
When using the b2b_bridge() function with the 'f' flag and provisional media,
we may end up in a state where the remaining entities in the call are
disconnected in terms of media, after the new entity rejects the call.
This commit will trigger an additional SDP renewal sequence between the
bridge initiator and it's original peer entity.