Migrate from index to references when pointing to a script route
The idx are hard to work with in the context of script reload, as they cannot be upated or checked (if still valid).
So, instead of pointing to the route by its idx, use an intermediary structure that holds: name of the route, the idx of the route, the version of the script.

Now, such intermediary structs (or route refs) may operate in 2 ways:
* they are per process, in pkg; all such refs are linked in a single list and automatically updated (their idx) upon reload.
* they are shared (like attached to data structs in shm), in shm; if so, before using it, the version is checked, and if old, the idx is updated on spot.