logging: improve the format of the logging interface's own errors
Align the format of the error logs from the consumers printing functions
with the rest of the plain stderror logs. For the event consumer, the
error logs can be sent as normally to the rest of the consumers
(stderror and syslog).

(cherry picked from commit 91c264abd63bfe200d0cf0f15aa300c5d1b4c799)