[uac_registrant] fix passing wrong context to authenticated REGISTER
When sending the authenticating REGISTER, be sure a new context is provided. This will avoid the usage of the context belonging to the first (challenged) REGISTER. This mixup of the context was breaking the tracing of these REGISTER requests via local route.

(cherry picked from commit c64766a8526c3953bd5adf45030a29a9aacb4550)