Get rid of outdated sched_yield.h temporary wrapper
The main utility of "sched_yield.h" was to provide sched_yield() on
OS'ed which didn't have it back in 2007, such as OpenBSD and NetBSD.

Nowadays, the landscape is different, with the issue having been fixed
in both above OS'es.  Moreover, by getting rid of the wrapper, we avoid
"missing -DHAVE_SCHED_YIELD" problems on architectures such as ARM (see
commit f8ad265a70f, where the issue was caused by this wrapper header
exclusively).