[b2b_logic] fix crash on tracing expired b2b tuples
When a tuple expires via timer, be sure to provide a processing context, otherwise the tracer (if enabled) will simply crash.
As part of the fix, two new helper functions were added for pushing and poping new processing contexts - these are to be reused by some prev similar fix on uac_registrant.
Related to #3110
Many thanks to @vtzan for helping with the troubleshooting and testing