Fix resuming partial msg reading on TCP
When conn rebalancing or parallel processing is enabled for TCP, a read operation may start (partial reading) in one process and be completed in a different one. So the con_req (buffer for partial reading) must be in shm mem and also it must survive a conn release (passing the conn back to MAIN).
Many thanks to @liviuchircu for reporting and helping with the troubleshooting
This issue was found during the TCP performance tests done for the 3.4 release