b2b_logic: properly reset the current tuple in all cases
The global variable used to identify the current tuple in the local_route was
not properly reset in all cases (when retrying a birdge action for example).
This could lead to a failure to identify the current tuple when acessing
the $b2b_logic variables, the b2b logic context etc., even when we're
not in the context of a local_route.

(cherry picked from commit 92772e69f36dcc70b2418374d83bf3274b13f104)