db_sqlite: fix possible crash when the values is relocated
When the query does not manage to count the rows in a first shot, the
values buffer might be reallocated, which can result in a relocation of
the buffer - if that's the case, we should repoint all the old values to
the new location, otherwise they will point to an invalid value.

Many thanks to Rob Moore and Steven Ayre from Dubber for reporting and
helping us testing and troubleshooting this corner case.

(cherry picked from commit 45aa9a947c6306f5723132f9c38903988be6cf8b)