rtp_relay: fix invalid contexts leaked in list
When a session would have been established on a provisional reply with
body (i.e. 183), the rtp contexts would have been stored in the list.
However, if the call would not establish (with a 200 OK), there would be
no one removing it from the list, even though the context was released -
this resulted in unexpected and almost impossible to track down invalid
memory usage.
The current patch makes sure that the context is removed from the list
if the call does not establish.

Many thanks to Rob Moore, Daniel Bryars and Steven Ayre from Dubber for
pushing an enormous amount of time into troubleshooting and debugging
this, as well as to Liviu Chircu for guiding the troubleshooting!

(cherry picked from commit d6ef279f63aa0682bf0fb7ef1d30a6abdfb8b942)