stir_shaken: Add the `e164_max_length` parameter
This parameter allows the 15-digit number length restriction of the E.164
format to be bypassed.  Especially useful in scenarios where various
telephony number prefixes are in use, causing some numbers to exceed
the standard maximum length.

Credits to @Integration-IT for watching over the core issue and helping
work towards the current solution!

Fixes #3202
Fixes #3182
Fixes #3181