tls_mgm: don't free anything in mod_destroy()
This fixes a possible double-free during shutdown.

The issue was that the `tls_mgm` module was unloaded (free'ing all of
its "domains") before the connections were destroyed. This left
connections with dangling pointers to the domains, and when the
connections are cleaned up, if the reference count is 0, `tls_mgm`
can then try to free the domain again, causing a crash.