aaa_diameter: Add management for "unreplied requests"
This becomes necessary both when the event_route is not defined, as well
as in various error cases, e.g. script writer omits to invoke
dm_send_answer() or an internal error occurs before the answer is built.

Finally, make sure to protect the requests list, since it is actually
managed by multiple threads part of the "Diameter peer" OpenSIPS worker:

* dm_update_unreplied_req() - called by the peer's Server Thread
    (i.e. "put a freshly received request on hold")
* dm_remove_unreplied_req() - called by the peer's Main Thread
    (i.e. "the script writer just built a new reply, msg can be freed")