Fix the E_CORE_SHM_THRESHOLD event
For some time now, this event has not been functional, due to
init_shm_mallocs() being called *before* parsing the opensips.cfg,
leading to un-initialized event holders and a quick-exit at runtime,
including times when the event actually needs to be raised.

Credits to Bogdan Iancu for reporting this issue!