cfg parser: Avoid unnecessary strlen on error case
When getline() returns -1, the @lineptr argument is never safe to read,
nor is this recommended.  So when both rc == -1 and EOF conditions
occur, it is safe to assume we read 0 bytes, without doing the strlen().

Many thanks to Dhiraj Mishra (@RandomDhiraj) for detecting, documenting
and reporting the potential risk of a heap buffer overflow here!