domain: init db connection in all processes
If the db_mode says we should cache the dialog table and only reload it
in MI proceeses, the module would not initialize the db connection for
SIP workers/timers. This would prevent the mi_script from running the
mi() commands.
This fix initializes the db connection in all proceses, thus enabling
the reload behavior from any process.