siprec: fix failover crash
When a failover happens, the initial_sdp is no longer available, hence
the module crashes when trying to use it. This fix makes sure that the
initial_sdp is not released, so that sequential failover can catch and
use them.

Many thanks to Farzaneh Soltanzadeh (@Fr-Soltanzadeh on GitHub) for
reporting it in #3421 and trying to fix it in #3425.
Fix #3421
Close #3425

(cherry picked from commit a970903c113b47c152c0077689c70faa2a1816cf)