sql_cacher: remove test key from cache
Since this key is stored in a non-base64 format, when the hash is being
iterated, this value breaks the convesion - and when trying to
`base64_decode("cdb_test_val")`, this results in a larger value than the
expected integer, resulting in a stack crash.

Removing the key from the hash fixes this crash, as all keys remain are
properly base64 encoded. Moreover, if the key is not properly encoded,
we should make sure that its decode does not overwrite the stack frame.

(cherry picked from commit dc38cf3a19819bc010489e0f88a663a4f35f9aca)