TCP: Fix race condition in "proto_tcp.tcp_parallel_handling"
During F_TCPCONN handling by a TCP worker, the @con pointer must *not*
be de-referenced if the READ was successful and connection already
returned to TCP Main.  This patch adds a new retcode to the tcp.read API
function in order to achieve this.

Credits to Yury Kirsanov and Bogdan Iancu for helping troubleshoot
this issue!