presence: Fix subtle race condition in update_db_subs()
While walking the list and collecting expired subscriptions, it is
incorrect to temporarily release, then re-acquire the lock and still
make use of the @s or @prev_s pointers afterwards, as the list may have
changed in the meantime!

The fix is to detach and collect any expired subs, then walk & handle
them in a separate iteration.