[I/O reactor] do sanity check on the provided fd index
On del operation, if an index (in the fd_array) is provided, check its consistency, to avoid corrupting the reactor's data.
Upon corruption, do a backtrace and force recovery by resetting the index to -1

IMPORTANT: this is not a fix, but a detection and recovery, hoping to help identifing bad ops over the reactor's fds.