tm: fix bogus contact built without domain
When fixing headers for fake messages (`fix_fake_req_headers` func), the
function was looking for lumps that were removing the contact, because
it would assume they will be the same that would add a new one (i.e.
`fix_nated_contact`). However, the `toppology_hiding("U")` function,
would create a del lump, but then adding 3 lumps over it. This would
make the fixing bogusly learning the contact as "sip:$user" instead of
hole URI.
The fix for this was to make sure that there is only one ADD lump,
otherwise it would point to a shorter (broken) URI.

Credits go to David Trihy from Genesys for reporting and helping us
troubleshoot.